Moved filter function
Filtering function moved from \controllers\index.js to \help\index.js .
