+++ backend/src/helper/index.js
@@ -11,7 +11,7 @@ export function filterObject(data, searchParameter, exclude){
     })
 }
 
-export function filterByFunction(data, searchParameter, relation){
+export function filterByRelation(data, searchParameter, relation){
     return JSON.parse(JSON.stringify(data)).filter(function (parameter) {
         return Object.keys(searchParameter).every(function (key) {
             return relation(parameter[key], searchParameter[key]);
@@ -27,7 +27,7 @@ export function mergeObject(data, searchParameter1, searchParameter2) {
 }
 
 export function fetchData(req) {
-    const data = req.query;
+    const data = req.body;
     return {
         fullName: data.fullname,
         age: data.age,
@@ -56,7 +56,7 @@ export function searchFor(user) {
         edu: {Category : 'University Degree'},
         tra: {Category : 'Training'},
         ctra: {Category: 'Certified Training'},
-        onsite: {'Mode of Delivery': "onsite"},
+        "onsite": {'Mode of Delivery': "onsite"},
         hybrid: {"Mode of Delivery": "hybrid"},
         jobLevel: {"Level": user.JobReadinessLevel},
         locLevel: {"local_lan_requirements": user.localLanguageLevel},
@@ -64,3 +64,50 @@ export function searchFor(user) {
         background: {"Cluster nb": user.background},
     }
 }
+
+
+export function filteredData(user, result) {
+
+    // First filter the location of the user - Global OR country
+    if(user.filledLocation)
+        result = mergeObject(result, searchFor(user).country, searchFor(user).global);
+
+    // Filter interest in entrepreneurship
+    if (user.filledEntrepreneur)
+        if(user.isEntrepreneur) {
+            result = filterObject(result, searchFor(user).enrepreneur);
+            return result; // Exit from function
+        } else {
+            result = filterObject(result, searchFor(user).enrepreneur, true);
+        }
+
+    // Split result to branches by category
+    let jobBranch = filterObject(result, searchFor(user).job);
+    let eduBranch = filterObject(result, searchFor(user).edu);
+    let trainBranch = filterObject(result, searchFor(user).tra);
+
+    // set mode of delivery, whether onsite or hybrid
+    jobBranch = mergeObject(jobBranch, searchFor(user).hybrid, searchFor(user).onsite);
+
+    // Filter, level <= applicant job readiness level
+    jobBranch = filterByRelation(jobBranch, searchFor(user).jobLevel, (dataVal, searchVal) => dataVal <= searchVal);
+
+    // Filter, language levels <= candidate language levels.
+    // English Language Levels
+    jobBranch = filterByRelation(jobBranch, searchFor(user).engLevel, (dataVal, searchVal) => dataVal <= searchVal);
+    // Local Language Levels
+    jobBranch = filterByRelation(jobBranch, searchFor(user).locLevel, (dataVal, searchVal) => dataVal <= searchVal);
+
+    // Filter which background matches their work background
+    jobBranch = filterObject(jobBranch, searchFor(user).background);
+
+    //if(jobBranch.length > 0)
+    result = jobBranch;
+    //else if(eduBranch.length > 0)
+    //  result = eduBranch;
+    //else
+    //  result = trainBranch;
+
+    return result;
+}
+
